rasa-dmt/requirements.txt
cryptography==43.0.1
numpy==1.23.5
peft
PyJWT==2.9.0
PyJWT==2.9.0
pymilvus
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
rasa-pro==3.10.3
rasa-sdk==3.10.0
Requests
sentence_transformers==3.0.1
torch==2.4.1

torchvision
accelerate
transformers
datasets
timm
einops
ftfy
tqdm
